### 2016-12-07 #132 @whedon
[PRE REVIEW]: Ramble: Higher-Order Functions for Parsing in R
**Submitting author:** @chappers (<a href="http://orcid.org/0000-0002-2089-3796">Chapman Siu</a>)
**Repository:** <a href="https://github.com/chappers/Ramble/tree/master/joss-paper" target ="_blank">https://github.com/chappers/Ramble/tree/master/joss-paper</a>
**Version:** v0.1.1
