### 2016-12-21 #132 @arfon
@noamross - would this also be a good candidate for rOpenSci?
