### 2016-12-26 #132 @pjotrp
Dear @chappers, before I look for a reviewer I would like to ask you why you think your project is relevant for science and do you have any applications in that area? Why is it not a proper R package with a vignette? I would like to read your PDF on slideshare, but to download it I need to register, something I will not do. Finally you'll need to make this FOSS with a proper license. Thanks.
