### 2016-12-26 #132 @chappers
Hi @pjotrp the details of the package and license can probably be easier viewed from the [CRAN link](https://cran.r-project.org/web/packages/Ramble/index.html), the software is already in a package form with license and [vignette](https://cran.r-project.org/web/packages/Ramble/vignettes/Higher-Order_Functions_for_Parsing_in_R.html). The slideshare file is irrelevant to the review.

The key application of this package is allow a user to write a parser in R which is capable of handling context-free grammars. Besides this package there does not appear to be any way to write a parser within R.

The applications could be in the data mining area when you have semi-structured data and are required to build a bespoke tool to parse strings into a data structure. It could also be used as a starting point for a compiler.
