### 2016-12-31 #132 @pjotrp
OK, my questions arise from reading the paper and the source tree. It will need to be clarified. All this information should be in either.

Do you know someone who could review your paper? Sorry for the delay, I was on a trip.
