### 2017-03-13 #132 @noamross
I can ping the rOpenSci crowd to see if someone is interested in reviewing this.
