### 2017-03-18 #132 @pjotrp
@arfon Since we are getting no pickup on this paper and I know something about R and parsers I'll take up review. I believe this is an exceptional situation.
