### 2017-03-25 #209 @pjotrp
Dear @chappers,

I finally have time to work on this review. I started, but there are two issues. First the license file does not contain an OSI approved license. This is obligatory for JOSS, so please add one. Then the paper you refer to is hosted on Slideshare https://www.slideshare.net/chapm0nsiu/ramble-introduction. I can not download it without having to register. Please make it freely available, perhaps as part of your repository.

Ping me when this is done. I will continue after.

Pj.
