### 2017-03-25 #209 @pjotrp
@chappers Also the unit tests point to a standard R test. Can you do better than that? Kinda meaningless to add a test icon to the website if it is not a real test. Correct me if I am wrong.
