### 2017-03-25 #209 @chappers
@pjotrp Updated the license with license.md (artifact of the R build process) and added the pdf to `/docs`.

What do you mean if its not a real test? Do you expect tests with sample data or something similar?
