### 2017-03-26 #209 @pjotrp
Thanks! Regarding the test I would expect it to build the package and run it as a minimum. Even better to have a real data test in it though. Many packaging systems depend on these tests to make sure integration works (including those of CRAN).
