### 2017-03-26 #209 @pjotrp
The current LICENSE file is a bit confusing. How about renaming LICENSE.md to LICENSE? Alternatively add that it is an MIT or expat license to the current LICENSE file.
