### 2017-03-26 #209 @pjotrp
If possible can you add to the README some guidelines on what people need to do to contact you? I.e., community guidelines.
