### 2017-03-26 #209 @pjotrp
I think this is a nice piece of work and can be quite useful to scientists needing to parse data. @chappers  to popularize this work I would add more examples of real use cases, especially for statistics or bioinformatics (who should be using proper parsers more often). Double them as tests for your test scripts too. Though you can do that later.

Before submission I recommend changing the name of the LICENSE file (also that allows github to show the license) and adding some community guidelines. Optionally you can improve paper.md a bit by referencing your PDF (now on github). If you feel energetic I would add a few use cases to paper.md too - just so as to make the reader aware of possibilities. When happy with your final version create a DOI for your source repository and ping @arfon for inclusion into JOSS.

Good job @chappers and sorry it took so long. The largest gap in time was after I asked you for potential reviewers and forgot to follow up when I got no response. Github issues do that do you: no message, no reminder. Next time I'll monitor more closely.
