### 2017-03-27 #209 @chappers
Hi @pjotrp thanks for spending time on this! I have updated the LICENSE file so now it shows up correctly. I have also added the contributing guide.

@arfon The software is in Zenodo: 10.5281/zenodo.438282
