### 2017-03-27 #209 @pjotrp
@arfon I believe this paper is good to be published in JOSS.
