### 2017-03-27 #209 @arfon
@chappers - the title of this paper appears to be ['Eclipse Golo'](https://github.com/chappers/Ramble/blob/master/joss-paper/paper.md) - is that correct?
