### 2017-03-27 #209 @arfon
@pjotrp many thanks for reviewing here ✨

@chappers - your paper is now accepted into JOSS and your paper DOI is http://dx.doi.org/10.21105/joss.00209 ⚡️ 🚀 💥
